requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.56
botocore==1.17.56
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-boto==0.3.12
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-filter==2.3.0
django-js-asset==1.2.2
django-storages==1.10
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
